% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shotmap_tdl.R
\name{create_shotmap_tdl}
\alias{create_shotmap_tdl}
\title{A plot function}
\usage{
create_shotmap_tdl(name_of_player = "Dusan Tadic", showstats = FALSE)
}
\arguments{
\item{name_of_player}{Name of the player defined in the excel file, default = Dusan Tadic.}

\item{showstats}{show the xG and means. default = FALSE}
}
\description{
This function creates a shotmap of a player for Tussen de Linies
}
\examples{
create_shotmap_tdl(name_of_player = "Dusan Tadic", showstats = FALSE)
ggsave("name.png",dpi=300,width=8,height=7,device=png)
}
\keyword{shotmap}
