% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shotmap_twitter.R
\name{create_shotmap_twitter}
\alias{create_shotmap_twitter}
\title{A plot function}
\usage{
create_shotmap_twitter(name_of_player = "Dusan Tadic", showstats = FALSE)
}
\arguments{
\item{name_of_player}{Name of the player defined in the excel file, default = Dusan Tadic.}

\item{showstats}{show the xG and means. default = FALSE}
}
\description{
This function allows you to express your love of cats.
}
\examples{
create_shotmap_twitter(name_of_player = "Dusan Tadic", showstats = FALSE)
}
\keyword{shotmap}
