% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEndKKD.R
\name{getEndKKD}
\alias{getEndKKD}
\title{Function that adds end locations
for xT}
\usage{
getEndKKD(With_name)
}
\arguments{
\item{df}{event data}
}
\description{
Function that adds end locations
for xT
}
\examples{
getEndKKD(df)
}
\keyword{crosses}
