% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXTPerCarry.R
\name{getXTPerCarry}
\alias{getXTPerCarry}
\title{Get xT values for carries only.
gives you the start, end en net xT value of a carry}
\usage{
getXTPerCarry(OneGame)
}
\arguments{
\item{dataframe}{Filter the data frame beforhand so it only contains successful carries with dribble_start_x,dribble_start_y,dribble_end_x & dribble_end_y}
}
\description{
Get xT values for carries only.
gives you the start, end en net xT value of a carry
}
\examples{
getXTPerPass(all_data)

}
\keyword{xThreat}
