% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXTPerPass.R
\name{getXTPerPass}
\alias{getXTPerPass}
\title{Get xT values
gives you the start, end en net xT value of a pass}
\usage{
getXTPerPass(OneGame)
}
\arguments{
\item{dataframe}{Filter the data frame beforhand so it only contains successful passes!}
}
\description{
Get xT values
gives you the start, end en net xT value of a pass
}
\examples{
getXTPerPass(all_data)

}
\keyword{xThreat}
