% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXTPerPassKKD.R
\name{getXTPerPassKKD}
\alias{getXTPerPassKKD}
\title{Get xT values
gives you the start, end en net xT value of a pass}
\usage{
getXTPerPassKKD(OneGame)
}
\arguments{
\item{dataframe}{Filter the data frame beforhand so it only contains successful passes!}
}
\description{
Get xT values
gives you the start, end en net xT value of a pass
}
\examples{
getXTPerPassKKD(all_data)

}
\keyword{xThreat}
