% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeOfPlay.R
\name{typeOfPlay}
\alias{typeOfPlay}
\title{A mutate function
Adds the type of play to a data frame with events}
\usage{
typeOfPlay(OneGame)
}
\arguments{
\item{OneGame}{df with data}
}
\description{
A mutate function
Adds the type of play to a data frame with events
}
\examples{
typeOfPlay(allGames)
}
\keyword{data}
\keyword{mutate}
